Refactored PowerShellTestsUtilities into a solution shared file.
