+++ src/Chapter10.Tests/Chapter10.Tests.csproj
@@ -4,6 +4,9 @@
   </PropertyGroup>
   <Import Project="..\ChapterTests.props" />
   <ItemGroup>
+    <Compile Include="..\Shared\Tests\PowerShellTestsUtilities.cs" Link="PowerShellTestsUtilities.cs" />
+  </ItemGroup>
+  <ItemGroup>
     <ProjectReference Include="..\Chapter10\Chapter10.csproj" />
   </ItemGroup>
   <ItemGroup>
