+++ src/Chapter10.Tests/Listing10.23.RegisteringAFinalizerWithProcessExit.Tests.cs
@@ -2,7 +2,8 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Diagnostics;
 using System.IO;
-using Chapter10.Tests.PowerShellTestsUtilities;
+using AddisonWesley.Michaelis.EssentialCSharp.Shared.Tests;
+
 using System.Text.RegularExpressions;
 
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter10.Listing10_23.Tests
@@ -10,32 +11,7 @@ namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter10.Listing10_23.Tests
     [TestClass]
     public class DisposeTests
     {
-
         static string Ps1Path { get; } = Path.GetFullPath("../../../../Chapter10/Listing10.23.RegisteringAFinalizerWithProcessExit.ps1", Environment.CurrentDirectory);
-        static bool PowerShellNotAvailable = PowerShellTestsUtilities.PowerShellNotInstalled;
-        static string PowershellEnvironmentVariableName { get; set; } = "powershell";
-
-        [ClassInitialize]
-        public static void ClassInitialize(TestContext testContext)
-        {
-            if (PowerShellNotAvailable) { Assert.Inconclusive("Powershell not installed"); return; }
-            
-            if (PowerShellTestsUtilities.WindowsEnvironment()) PowershellEnvironmentVariableName = "powershell";
-            else PowershellEnvironmentVariableName = "pwsh";
-            
-            string testStatus = "create";
-            Process powershell = Process.Start(PowershellEnvironmentVariableName, $"-noprofile -command \"{Ps1Path} 0 null {testStatus}\"");
-            powershell.WaitForExit();
-        }
-
-        [ClassCleanup]
-        public static void RemoveProcessExitProj()
-        {
-            if (PowerShellNotAvailable) { return; }
-            string testStatus = "cleanup";
-            Process powershell = Process.Start(PowershellEnvironmentVariableName, $"-noprofile -command \"{Ps1Path} 0 null {testStatus}\"");
-            powershell.WaitForExit();
-        }
 
         [DataTestMethod]
         [DataRow("processExit", FinalizerRegisteredWithProcessExit, DisplayName = "Finalizer Registered With ProcessExit")]
@@ -43,8 +19,6 @@ namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter10.Listing10_23.Tests
         [DataRow("gc", GCCalled, DisplayName = "Garbage Collected called")]
         public void FinalizerRunsAsPredicted_ConsoleOutputIsInOrder(string finalizerOrderOption, string expectedOutput)
         {
-            if (PowerShellNotAvailable) { Assert.Inconclusive("Powershell not installed"); return; }
-
             string traceValue = "0";
             string testStatus = "run";
